Initial Commit
This commit has the main structure that I think that I am going to use.
Game being the main driving struct.

Each piece in pieces having to implement the trait PieceMove is what I'm
running with to iterate over every piece to check for checks.

I have the tests for the movement, but there is still confusion on if
is_valid like function will be responsible for captures or the can_move.

I'm not sure about what cargo fmt did to the tests.  I'm not sure if I
should change the way I write them.  Mainly the assert messages