initial parse game function is create
I don't really have any tests for this and it's kind of a pain in the
butt.

What I'm going to do is rely on integration testing where I have a few
games that I have painstakingly written out as tests.

Then loop through a whole bunch of games and make sure that none of them
panic or assert.

One part I know I need to add is the parsing of the result.  That is
1-0, 1/2-1/2, and 0-1