CHES-16 add play_move
There is now a function that takes in a parsed move and plays the move.

I do have 2 games that run where I have manually checked the end
position, but I don't have any automated checks on them.
I also need to add checks, checkmate, and stalemates to the game