created a gui in the gui_runner module
This is a pretty simple gui.  It shows the board and then still uses
text as the moves.  There is no mouse functionality, but it could be
added in the future.  I also found that there may be some bugs
with pawn movement.

I also recognize that the function to create the board container is
not good at all.  I believe this would be a good place for a macro,
but I haven't read about them yet and so I settled on a gross function
to get it working